[Refactor] Implement AbstractCommands base class
**Major**:
- Created AbstractCommands abstract base class with shared functionality for command handlers
- Refactored BaseChatCommands to inherit from AbstractCommands instead of standalone implementation
- Refactored HatchCommands to inherit from AbstractCommands instead of standalone implementation
- Centralized common initialization logic (chat_session, settings, env_manager, logger) in AbstractCommands

**Minor**:
- Moved _parse_args() method from individual classes to shared AbstractCommands base
- Moved _print_command_help() method from individual classes to shared AbstractCommands base
- Moved get_command_metadata() method from individual classes to shared AbstractCommands base
